# openapi-spec-formatted.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Healthcare System API",
        "version": "1.0.0",
        "description": "API documentation for Smart Healthcare System - UTS IAE/Web Service Development",
        "contact": {
            "name": "Healthcare Team"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server (API Gateway)"
        }
    ],
...
    "paths": {
...
        "/api/doctors": {
            "get": {
                "summary": "Get all doctors",
                "tags": [
                    "Doctors"
                ],
                "responses": {
                    "200": {
                        "description": "List of all doctors"
                    }
                }
            },
            "post": {
                "summary": "Create new doctor",
                "tags": [
                    "Doctors"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "specialization",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. Jane Smith, Sp.PD"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "example": "Cardiologist"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567891"
                                    },
                                    "schedule": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Monday 09:00-12:00",
                                            "Wednesday 14:00-17:00"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Doctor created successfully"
                    }
                }
            }
        },
        "/api/doctors/{id}": {
            "get": {
                "summary": "Get doctor by ID",
                "tags": [
                    "Doctors"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Doctor ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor details"
                    },
                    "404": {
                        "description": "Doctor not found"
                    }
                }
            },
            "put": {
                "summary": "Update doctor",
                "tags": [
                    "Doctors"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doctor updated successfully"
                    }
                }
            },
            "delete": {
                "summary": "Delete doctor",
                "tags": [
                    "Doctors"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor deleted successfully"
                    }
                }
            }
        },
        "/api/appointments": {
            "get": {
                "summary": "Get all appointments",
                "tags": [
                    "Appointments"
                ],
                "responses": {
                    "200": {
                        "description": "List of all appointments"
                    }
                }
            },
            "post": {
                "summary": "Create new appointment",
                "description": "Creates a new appointment with validation against Patient and Doctor services",
                "tags": [
                    "Appointments"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "patient_id",
                                    "doctor_id",
                                    "appointment_date"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "string",
                                        "example": "507f1f77bcf86cd799439011"
                                    },
                                    "doctor_id": {
                                        "type": "string",
                                        "example": "507f1f77bcf86cd799439012"
                                    },
                                    "appointment_date": {
                                        "type": "string",
                                        "example": "2024-11-15 10:00"
                                    },
                                    "complaint": {
                                        "type": "string",
                                        "example": "Chest pain and shortness of breath"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully"
                    },
                    "404": {
                        "description": "Patient or Doctor not found"
                    }
                }
            }
        },
        "/api/appointments/{id}": {
            "get": {
                "summary": "Get appointment by ID",
                "tags": [
                    "Appointments"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment details"
                    }
                }
            },
            "delete": {
                "summary": "Delete appointment",
                "tags": [
                    "Appointments"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment deleted successfully"
                    }
                }
            }
        },
        "/api/appointments/{id}/status": {
            "put": {
                "summary": "Update appointment status",
                "tags": [
                    "Appointments"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "confirmed"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    }
                }
            }
        },
...
    }
}